Auth and Guard done 🍭
Sign Up, Sign In, Sign Out and the guard that checks the token that comes in the headers already work. Just as I made the user go online at the time of login and after logging out, check if the user is already offline or if he is still online and the result is applied to his due logic.