Info Modal 😷
Modal component has been created to view all data related to the anime.

We just need to create the comments component and add the logic. If the anime has more than 3 reviews, it will open a small modal with all the reviews as if it were a forum thread (but much simpler.)