Refactor and New features 🤘
The Card component and the modal that shows the information of the anime, were subdivided into several components to facilitate the reading and maintenance of the code.

In the same way that color mode was added to the application, users already have the button to switch between dark or light mode.

As well as the modal was added to view all the reviews that the featured anime holds.