Small changes 🤙
I'm using the Chakra UI bookstore component called "Editable" but I still don't know how to send the most recent data to the backend.

A json file called dummydb was added to have an exemplary array of what the backend response would be like.

The Review Detail component has been updated.

A Button Group has been added to identify the username that has the session started and the consequent button to end the session.