Client State ✊
The store and store changes were added to the project.

Sign Up, Sign In and Log Out are already in operation.

In the same way that it is already possible to go, I acquired the value of the id of the user who is logged in.

In addition to all this, an instance of axios was created.